Establish core-owned Runner API
