+++ crates/contracts/homeboy-runner-contract/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "homeboy-runner-contract"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+description = "Transport-neutral runner execution contracts for Homeboy"
+repository = "https://github.com/Extra-Chill/homeboy"
+publish = false
+
+[dependencies]
+homeboy-api-jobs-contract = { version = "0.1.0", path = "../homeboy-api-jobs-contract" }
+homeboy-lab-contract = { version = "0.1.0", path = "../homeboy-lab-contract" }
+homeboy-source-snapshot-contract = { version = "0.1.0", path = "../homeboy-source-snapshot-contract" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
