+++ crates/homeboy-core/Cargo.toml
@@ -32,6 +32,7 @@ homeboy-finding = { version = "0.1.0", path = "../homeboy-finding" }
 homeboy-paths = { version = "0.1.0", path = "../homeboy-paths" }
 homeboy-product-identity = { version = "0.1.0", path = "../homeboy-product-identity" }
 homeboy-release-contract = { version = "0.1.0", path = "../contracts/homeboy-release-contract" }
+homeboy-runner-contract = { version = "0.1.0", path = "../contracts/homeboy-runner-contract" }
 homeboy-audit-contract = { version = "0.1.0", path = "../contracts/homeboy-audit-contract" }
 homeboy-refactor-contract = { version = "0.1.0", path = "../contracts/homeboy-refactor-contract" }
 homeboy-code-audit = { version = "0.1.0", path = "../homeboy-code-audit" }
